Normalize Gemfile & pin tests to rspec < 2.99
Pin to rspec < 2.99 until rspec-puppet officially supports rspec 3.x.
This is to avoid warnings due to deprecated matchers in the
rspec-puppet gem.

This change is also intended to normalize the Gemfile for all stackforge
puppet modules.  This should facilitate pinning gem versions when
releasing new stable branches, since the same diff can be used for all
repos.

This also removes the Gemfile.lock, since this is the only stackforge
puppet module that has one.

Change-Id: I2f55a7a5339c177e3c1d19a9e165b5c5fe602d4c
Partial-Bug: #1326034