Update stdlib dependency to 4.x
ironic's dependency on stdlib 2.3.0 is out of sync with the other
openstack modules. This is a problem because it is not possible to
install this module alongside the other openstack modules because of the
dependency conflicts. This patch fixes the issue by updating the
depenency on stdlib to 4.x. The stdlib functions currently in use in
ironic have carried on through 4.x.

Change-Id: I8d8f11049587b0d1e5a66914224d7ed922e51c4d