Run dbsync as root and don't manage /etc/ironic & ironic.conf ownership
Ubuntu packaging is running dbsync command as root during ironic-common
postinstall script so when Puppet tries to run dbsync again, it fails
because it is run with ironic user.
So we need to run dbsync as root for now.
This is a temporary patch until it's changed in Packaging.

Also, don't manage user /etc/ironic & ironic.conf ownership and chmod to be packaging agnostic
(ubuntu & fedora use different permissions) and allow Puppet
idempotency.

Change-Id: Ia3571e4a212d9493604494bce5592da0831d9d7a
Partial-bug: #1450942