keystone/auth: make service description configurable
This commit adds the service description as a class parameter in order to allow
users to update from a previous version if the service description is changed
(incorrectly spelled or wrong description).

This commit also fix the Ironic service description, 'Ironic Networking Service'
seems to be a too quick sed of a neutron manifest.

Closes-Bug: #1468407
Change-Id: I8ad187ae171b6e6c9841b18012a04a504ca75c5f