Creation of ironic::db::sync
In order to standardize the way dbsync are run across our modules,
we create a new class ironic::db::sync.
This class will be included if sync_db is enabled.

By making this transition the ironic::db::sync class
can be returned by the ENC.

A use case would be in an highly available environment, with 3 galera
nodes, include ironic on every node with sync_db set to false
and have the ENC return ironic::db::sync just for one node.

Change-Id: I79829a1744a98f3c9b8ad6f71fd2f5e5924e463e