[inspector] Specify SELinux type context for tftp dir
We create a /tftpboot/ dir in the root directory without
specifying a SELinux type context, so it gets default_t. The
tftp process is not allowed read permission on default_t, and
as a result the ironic-inspector-dnsmasq service fails to
start.

The fix is to specify a type context of tftpdir_t when
creating the /tftpboot/ directory.

This patch also pulls the /tftpboot/ directory out of the if
block relating only to tftp configuration. The /tftpboot/
directory is also used by iPXE/http for the undionly.kpxe
chainloading image.

Change-Id: Id20802b457a8c75d8b1af04a2d2f56b6b776536e
Closes-Bug: 1515630
(cherry picked from commit d9611fa)