Do not use notify for db-sync from main init class
If we want to use db-sync only on some node (for example, primary node)
and don't want run it on another we will get resource not found error
for notify in package ironic-common on db-sync.

Also db/sync class already contains all needed notifies on ironic-package
tag, so it also is duplicate.

Change-Id: I29939384bf96b82d9124307222a239b8aa95336b
Related-bug: #1529046