Fix duplicate parameters and drop mock
This change removes the duplicate parameters from the main ironic class.
With the update to 3.8.5, this now shows up as a syntax error in the
syntax jobs. We should only define these parameters once.

Also mock is not used anywhere in the module, let's drop it.
This patch will help to hit this bug:
https://launchpad.net/bugs/1492636

Change-Id: Iaa6acafdc771118c74f85f3332f01b58d233743b
Closes-Bug: #1537251
Co-Authored-By: Emilien Macchi <emilien@redhat.com>