[inspector] use dnsmasq dhcp-sequential-ip flag in dnsmasq.conf
Introspection naturally happens in large bulks, after which it's inactive.
Small pool for DHCP addresses means that we'll have conflicts due to how
dnsmasq distributes them by default - using hashing. This change tells dnsmasq
to allocate IP addresses sequentially instead to avoid these conflicts.

The drawback of this option is that long-running clients may switch IP
addresses if their lease expires. This is not a concern for short introspection
process.

Change-Id: I0f08609a9f72799ef9f62216041f2b2b9795afd5