Add transport_url parameter for oslo.messaging
This commit adds the transport_url parameter for oslo.messaging. The
url is of the form:

  transport://user:pass@host1:port[,hostN:portN]/virtual_host

Where the transport scheme specifies the rpc or notification backend
as one of rabbit, amqp, zmq, etc. Oslo.messaging is deprecating the
host, port, and auth configuration options [1]. All drivers will get
these options via the transport_url.

This patch:
* use oslo::messaging::default resource
* add parameter for transport_url
* update spec test
* add feature release note

[1] https://review.openstack.org/#/c/317285/

Change-Id: I29525aabbea99e36546643f4e80d4970fc2a04ac