Remove "dhcp" command from the iPXE script
At this point we already have DHCP on the right NIC to be able to download
this iPXE script. The "dhcp" may actually break the boot, as it runs DHCP
on the first available interface, not the PXE booting one.

Change-Id: I9ec62b6b662c9ea70f7cc12bd0567b5e4119faf7
Closes-Bug: #1635191
(cherry picked from commit 806d5a15693f7a0ce0cd379e7af81275bf334982)