Increase Ironic inspector dnsmasq lease time
As described in bug, its possible that addresses provided to IPA via
DHCP may expire and not be added to the introspection report.  This
increases the dhcp lease time from 2 to 10 minutes.  There should be
no downside to increasing this lease time since these addresses
are only used during introspection.

Change-Id: Iae15e9db0acc2ecd5b087a9ca430be948bc3e649
Closes-Bug: 1721051