Deprecate and unset the enabled_drivers option
The classic drivers are being removed, with the removal of pxe_ipmitool
already planned. The enabled_drivers option is deprecated in this change
and is made no-op, since there are no more classic drivers to enable
with it. We also force it to be empty, since the ironic team is planning
on a validate that is not set on ironic-conductor start up.

Change-Id: I6e02f59807c98643923d1c2333fcb25114735f25