Correct default number of threads
The default number of threads does not match the default
tested upstream[1], which can return errors for users when
requests begin backing up in the webserver due to a lack
of resources and the nature of the ironic API having some
synchronous requests.

Upstream ironic, as of the Rocky cycle, executed with uwsgi
which is configured for the same process count algorithm,
but enable threds. There does not appear to be an upper bound
to that setting, but accoriding to mod_wsgi documentation[3],
the default is 15 threads. The example shipped with ironic[4]
defaults to 10, which also seems to be a valid configuration.

[1]: https://git.openstack.org/cgit/openstack/ironic/tree/devstack/lib/ironic?h=stable/rocky#n1129
[2]: http://git.openstack.org/cgit/openstack-dev/devstack/tree/lib/apache#n267
[3]: https://modwsgi.readthedocs.io/en/develop/configuration-directives/WSGIDaemonProcess.html
[4]: https://git.openstack.org/cgit/openstack/ironic/tree/etc/apache2/ironic?h=stable/rocky#n21

Story: #2003464
Task: #24724
rhbz: #1613995

Change-Id: I327219f96cfc0a69643ab6369cb8e831251f799c
(cherry picked from commit 037cd2cb5d332ba178d78390a3f5bcd75ce1e9e3)