Expose dnsmasq_pxe_filter/purge_dhcp_hostsdir opt
The purge_dhcp_hostsdir controls wheater Ironic will
purge the dhcp_hostdir on initialization.

When the directory is purged the dnsmasq service must
be restarted as well to ensure the DHCP server config
is in sync once the dhcp-hosts directory is populated
with new content. Depending on how services are
deployed, controlling the dnsmasq service is non-
trivial. (I.e containers or different hosts.)

Change-Id: Icc532115891c567dde20a28110bf08f54187c49f
Related-Bug: #1780421