Clean up ipxe support
The puppet-ironic module's ipxe support does not align with current
versions of ironic which delineated ipxe and pxe support quite some
time ago.

This is an attempt to reconcile this and also change the default
ipxe binary for EFI from ipxe.efi which is intended to contain
firmware, and snponly.efi which only has the EFI networking stack.
In the vast majority of cases, short of very broken EFI firmware,
operators need snponly.efi.

Change-Id: Ia30aff290ec24972f387612851f8f630ddc9403b