Fix ppc64le support to coexist with PXE preference
In essence, the pxe preference doesn't apply to ppc64le.

This is because PXE data loading is part of the firmware and
does not use an intermediate boot loader like ipxe, syslinux,
or even grub.

Changes the logic to always set the stage for PXE based
deployments and enables boot setting awareness for ppc64le
based on the TFTP defaults. This allows ppc64le to coexist
with an environment preferring HTTP transport like those
used with iPXE.

Change-Id: I27fa53700c1b44f14725bc66b929732553da58f8