Deprecate bifrost support
My downstream team does not maintain this and the settings
which are embedded and encoded in puppet-ironic are incorrect
and no longer match the current state of Bifrost as an independent
sub-project of ironic.

Given this state, and the fact it is largely independent tooling
which duplicates much of puppet-ironic, it does not make sense to
maintain, and is thus deprecated.

Change-Id: Ia456fcea46c44c23d8c99fbf3bb59b63180a28f2