Use native implementation to add listen for apache vhost
Currently the vhost management in ironic::pxe has its own logic to
inject the Listen option into the vhost configuration file instead of
the core apache configuration file but these two should be effectively
same as long as vhosts are properly configured.

Because we globally use the implementation in puppetlabs-apache which
adds Listen to the core file, this replaces own mechanism by that
existing implementation to simplify the own logics.

Change-Id: I1e979159116f25a404b5f081c02352a6f2412074