Debian/Ubuntu: Fix installation of pxelinux.0 and syslinux files
Debian and Ubuntu requires the separate pxelinux package to use
the pxelinux.0 file and the file is stored in a separated directory.

Also, the current directory to locate the syslinux files are outdated
and does not work with the recent versions.

This change fixes these two problems.

Conflicts:
	manifests/params.pp
	spec/classes/ironic_pxe_spec.rb

Change-Id: I560b3b6af5b2bab0b1144aa1eda6088f01336d75
(cherry picked from commit b780aca6499a7504ef72195eb4243801d53ea5aa)