Set [deploy] http_root option using the common class input
Previously the [deploy] http_root option is set independently from
the common pxe parameters but this requires that users set both of
these options consistently.

This ensures the option is set according to the common definition.

This also ensures both http_root and tftp_root have the hard-coded
default value, because we rely on these values heavily in setting up
tftp server and http server and any change in ironic side can break
the deployment directly.

Change-Id: Ia26fa37b3fcba7d89c3d39c7d91b8c6e4cffde8c