inspector: Show warning when dhcp hostsdir is not configured
The dnsmasq_dhcp_hostsdir should be set to a specific path when dnsmasq
PXE filter is used. This adds a warning message to inform possible
mis-configuration.

Note that dhcp hostsdir is not needed when the other PXE filter such
as iptables PXE filter is used, thus we avoid updating the default
value.

Closes-Bug: #2042731
Change-Id: If127bd4661709835ab0d9d39e0341367fdef51b7