Use new openstackclient tag
The new openstackclient tag was added so that we can get all resources
about openstack CLI more easily. This adds this tag to ironicclient
because the plugin provides some sub-commands.

This also removes the support package tag because the client package is
not at all used by ironic itself.

In addition, openstackclient is needed to query keystone api when
transform in ironic_config is used.

Depends-on: https://review.opendev.org/899594
Change-Id: I46dcf0f8e7f9b72d27317d310bffdd25653a25cf